Fix XDT transforms failing on Mono.
Source code fixes:
https://github.com/mrward/xdt/commit/b2c3b5383d589c3f79650a0cab93f88a8741b057
https://github.com/mrward/xdt/commit/9ae2e113dc4140fa7da853436f42547795cfebb5

Fix Remove and SetAttributes transforms on Mono.

Running an xml transform using xdt:Transform='Remove' would fail
on Mono with a null reference exception.

The XmlElementContext TargetNode would be null in Mono resulting
in a null reference when the name of the node was logged. This
would only occur if the original xml node existed.

Microsoft.Web.XmlTransform.XmlNodeException: Object reference not set to an instance of an object ---> System.NullReferenceException: Object reference not set to an instance of an object
  at Microsoft.Web.XmlTransform.Remove.RemoveNode ()
  at Microsoft.Web.XmlTransform.Remove.Apply ()
  at Microsoft.Web.XmlTransform.Transform.ApplyOnce ()
  at Microsoft.Web.XmlTransform.Transform.Execute (Microsoft.Web.XmlTransform.XmlElementContext context, System.String argumentString)
  --- End of inner exception stack trace ---
  at Microsoft.Web.XmlTransform.XmlTransformationLogger.LogErrorFromException (System.Exception ex)
  at Microsoft.Web.XmlTransform.Transform.Execute (Microsoft.Web.XmlTransform.XmlElementContext context, System.String argumentString)
  at Microsoft.Web.XmlTransform.XmlTransformation.HandleElement (Microsoft.Web.XmlTransform.XmlElementContext context)

Using an xml transform with xdt:Transform="SetAttributes" that updates
attributes fails when running on Mono with an Argument cannot be null
exception.

Use custom build of the Microsoft.Web.XmlTransform assembly that
sets the prefix on XmlAttributes to an empty string instead of
null.