Support running ScriptCS scripts.
When a package is installed the following scripts are run:

    init.csx
    install.csx

When a package is uninstalled the following script is run:

    uninstall.csx

The ScriptCS script can interact with the IDE via global properties.
InstallPath, ToolsPath, Package, Project and DTE are now available
as global properties. Logging methods are available as global methods.
Host information is now available from a ScriptHost object.

Example script below:

    Log ("ScriptHost.Name: " + ScriptHost.Name);
    Log ("ScriptHost.Version: " + ScriptHost.Version);
    Log ("NuGet version: " + ScriptHost.NuGetVersion.ToString());
    Log ("InstallPath: " + InstallPath);
    Log ("ToolsPath: " + ToolsPath);

    if (Package != null) {
    	Log ("Package.ToString (): " + Package.ToString ());
    }

    Log ("DTE.Version: " + DTE.Version);
    if (Project != null) {
    	Log ("Project.Name: " + Project.Name);
        DTE.ItemOperations.Navigate ("http://xamarin.com");
    }