Allow Microsoft.ApplicationInsights to be installed with NuGet.Core.
Fixed bug #30049 - Unable to install Microsoft.ApplicationInsights
https://bugzilla.xamarin.com/show_bug.cgi?id=30049

The Microsoft.ApplicationInsights NuGet package specifies a
minClientVersion of 2.8.50313 which is the version of the NuGet
Package Manager's assembly version released by Microsoft but not the
production version which is 2.8.1. NuGet.Core when built for
MonoDevelop uses the same version for both of these values and does
not match the version generated by Microsoft.

The Microsoft.ApplicationInsights NuGet package seems to be the only
NuGet package that uses the assembly version. Other NuGet packages
produced by teams at Microsoft use the product version when specifying
the minClientVersion.

To handle this the 2.8.50313 version is mapped to 2.8.1. Another
alternative fix would have been to change the assembly version to
match the one generated by Microsoft. However most of the existing
NuGet packages are incorrectly using the product version as the
minClientVersion, so they would install into project when the NuGet
product version was not high enough since the check is made against
the assembly version. This problem only seems to have started since
NuGet started using 3 numbers for the 2.8.* releases. The third
number in the assembly version is based on the date the assembly
was built.